% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{emp_cdf}
\alias{emp_cdf}
\title{Empirical distribution functions}
\usage{
emp_cdf(x, x_f, x_e, eps)
}
\arguments{
\item{x}{a vectors of values}

\item{x_f}{lower bound of x}

\item{x_e}{upper bound of x}

\item{eps}{tolarance}

\item{...}{arguments for densratio::}
}
\value{
function of empirical pdf/cdf
}
\description{
Empirical cdf with cumsum. Empirical pdf with kernel density function
}
\examples{
x <- rnorm(100)
cdf_x <- emp_cdf(x)
print(cdf_x(0))
}
